+++ src/app/errors/ValidationError.ts
@@ -1,0 +1,8 @@
+export default class ValidationError extends Error {
+
+	constructor(message: string) {
+		super(message);
+
+		Object.setPrototypeOf(this, ValidationError.prototype);
+	}
+}