+++ src/app/types/bounty/BountyReward.ts
@@ -1,0 +1,5 @@
+export type BountyReward = {
+	amount: number,
+	currencySymbol: string,
+	scale: number
+};